Add files via upload
DISCLAIMER:
no scientific reference is to be assumed for any usage of this sheet. 
this is purely a personal and non-professional attempt of analyzing this event.

DESCRIPTION:
a. data collection
the analysis of the events is taken into consideration via a (present day):(previous day) ratio calculation.
no single and different sub-events rather than the two polarized macro-categories (which sums up any possible incomes/outcomes) or additional information about the containment of the disease are concerned
Each arrow symbolizes a (healthly) positive or negative advancing of the spread in comparison of the overall data

b. forecasting
this rawly and roughly utilizes the median ratio of the last 7 days to project a forecasting to the hopefully expected date where the two sums will cancel each other out.